requirements.txt
moderngl
numpy
PyGLM
pywavefront
cython
pillow
setuptools
glfw
imgui
pyopengl
pyyaml